requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
httpx==0.28.1
numpy==1.26.4
faiss-cpu==1.7.4
sentence-transformers==3.0.1
python-telegram-bot==22.2
python-dotenv==1.1.1
pytest==8.4.1
